refactor(botonic-react): create use-notifications hook to resolve all enabled attributs and custom components used in notifications
